microservices/school-service/requirements.txt
uvicorn
fastapi
httpx
python-consul


